Use const SigMatchCtx* for Match() arguments
The SigMatch contex should not be changed by the Match() funciton, so pass it
as a const SigMatchCtx*.

Straight forward changes except for DetectFlowintMatch() was setting
target.tvar.idx inside the DetectFlowintData cast from const SigMatchCtx *.
Instead use a local variable.

The only other place that uses target.tvar.idx was in a debug message, so it
was removed.

In fact, target.tvar.idx is next assigned, so it could be removed.